+++ examples/cc_ha/main.tf
@@ -95,10 +95,25 @@ resource "local_file" "user_data_file" {
   filename = "../user_data"
 }
 
+################################################################################
+# Locate Latest CC AMI by product code
+################################################################################
+data "aws_ami" "cloudconnector" {
+  most_recent = true
+
+  filter {
+    name   = "product-code"
+    values = ["2l8tfysndbav4tv2nfjwak3cu"]
+  }
+
+  owners = ["aws-marketplace"]
+}
+
 # Create specified number of CC appliances
 module "cc_vm" {
   source                    = "../../modules/terraform-zscc-ccvm-aws"
   cc_count                  = var.cc_count
+  ami_id                    = contains(var.ami_id, "") ? [data.aws_ami.cloudconnector.id] : var.ami_id
   name_prefix               = var.name_prefix
   resource_tag              = random_string.suffix.result
   global_tags               = local.global_tags
@@ -132,6 +147,7 @@ module "cc_iam" {
   resource_tag        = random_string.suffix.result
   global_tags         = local.global_tags
   cc_callhome_enabled = var.cc_callhome_enabled
+  secret_name         = var.secret_name
 
   byo_iam = var.byo_iam
   # optional inputs. only required if byo_iam set to true