+++ modules/terraform-zscc-ccvm-aws/main.tf
@@ -12,26 +12,11 @@ EOF
 
 
 ################################################################################
-# Locate Latest CC AMI by product code
-################################################################################
-data "aws_ami" "cloudconnector" {
-  most_recent = true
-
-  filter {
-    name   = "product-code"
-    values = ["2l8tfysndbav4tv2nfjwak3cu"]
-  }
-
-  owners = ["aws-marketplace"]
-}
-
-
-################################################################################
 # Create Cloud Connector VM
 ################################################################################
 resource "aws_instance" "cc_vm" {
   count                       = local.valid_cc_create ? var.cc_count : 0
-  ami                         = data.aws_ami.cloudconnector.id
+  ami                         = element(var.ami_id, count.index)
   instance_type               = var.ccvm_instance_type
   iam_instance_profile        = element(var.iam_instance_profile, count.index)
   vpc_security_group_ids      = [element(var.mgmt_security_group_id, count.index)]
@@ -39,6 +24,11 @@ resource "aws_instance" "cc_vm" {
   key_name                    = var.instance_key
   associate_public_ip_address = false
   user_data                   = base64encode(var.user_data)
+
+  metadata_options {
+    http_endpoint = "enabled"
+    http_tokens   = var.imdsv2_enabled ? "required" : "optional"
+  }
 
   tags = merge(var.global_tags,
     { Name = "${var.name_prefix}-cc-vm-${count.index + 1}-${var.resource_tag}" }
@@ -67,12 +57,6 @@ resource "aws_network_interface" "cc_vm_nic_index_1" {
   )
 }
 
-# Get Data info of NIC to be able to output private IP values
-data "aws_network_interface" "cc_vm_nic_index_1_eni" {
-  count = local.valid_cc_create ? var.cc_count : 0
-  id    = element(aws_network_interface.cc_vm_nic_index_1[*].id, count.index)
-}
-
 
 ################################################################################
 # Create Cloud Connector Service Interface #1 for Medium/Large CC. 
@@ -92,12 +76,6 @@ resource "aws_network_interface" "cc_vm_nic_index_2" {
   tags = merge(var.global_tags,
     { Name = "${var.name_prefix}-cc-vm-${count.index + 1}-${var.resource_tag}-SrvcIF-2" }
   )
-}
-
-# Get Data info of NIC to be able to output private IP values
-data "aws_network_interface" "cc_vm_nic_index_2_eni" {
-  count = local.valid_cc_create && var.cc_instance_size != "small" ? var.cc_count : 0
-  id    = element(aws_network_interface.cc_vm_nic_index_2[*].id, count.index)
 }
 
 
@@ -121,12 +99,6 @@ resource "aws_network_interface" "cc_vm_nic_index_3" {
   )
 }
 
-# Get Data info of NIC to be able to output private IP values
-data "aws_network_interface" "cc_vm_nic_index_3_eni" {
-  count = local.valid_cc_create && var.cc_instance_size != "small" ? var.cc_count : 0
-  id    = element(aws_network_interface.cc_vm_nic_index_3[*].id, count.index)
-}
-
 
 ################################################################################
 # Create Cloud Connector Service Interface #3 for Large CC. This resource will 
@@ -146,10 +118,4 @@ resource "aws_network_interface" "cc_vm_nic_index_4" {
   tags = merge(var.global_tags,
     { Name = "${var.name_prefix}-cc-vm-${count.index + 1}-${var.resource_tag}-SrvcIF-4" }
   )
-}
-
-# Get Data info of NIC to be able to output private IP values
-data "aws_network_interface" "cc_vm_nic_index_4_eni" {
-  count = local.valid_cc_create && var.cc_instance_size == "large" ? var.cc_count : 0
-  id    = element(aws_network_interface.cc_vm_nic_index_4[*].id, count.index)
 }