+++ modules/terraform-zscc-ccvm-aws/outputs.tf
@@ -35,20 +35,20 @@ output "id" {
 
 output "cc_service_private_ip" {
   description = "Instance Device Index 1 Private IP"
-  value       = data.aws_network_interface.cc_vm_nic_index_1_eni[*].private_ip
+  value       = aws_network_interface.cc_vm_nic_index_1[*].private_ip
 }
 
 output "cc_med_lrg_service_1_private_ip" {
   description = "Instance Device Index 2 Private IP"
-  value       = data.aws_network_interface.cc_vm_nic_index_2_eni[*].private_ip
+  value       = aws_network_interface.cc_vm_nic_index_2[*].private_ip
 }
 
 output "cc_med_lrg_service_2_private_ip" {
   description = "Instance Device Index 3 Private IP"
-  value       = data.aws_network_interface.cc_vm_nic_index_3_eni[*].private_ip
+  value       = aws_network_interface.cc_vm_nic_index_3[*].private_ip
 }
 
 output "cc_lrg_service_3_private_ip" {
   description = "Instance Device Index 4 Private IP"
-  value       = data.aws_network_interface.cc_vm_nic_index_4_eni[*].private_ip
+  value       = aws_network_interface.cc_vm_nic_index_4[*].private_ip
 }