+++ modules/terraform-zscc-gwlb-aws/main.tf
@@ -2,11 +2,12 @@
 # Configure target group
 ################################################################################
 resource "aws_lb_target_group" "gwlb_target_group" {
-  name        = "${var.name_prefix}-cc-target-${var.resource_tag}"
-  port        = 6081
-  protocol    = "GENEVE"
-  vpc_id      = var.vpc_id
-  target_type = "ip"
+  name                 = var.target_group_name
+  port                 = 6081
+  protocol             = "GENEVE"
+  vpc_id               = var.vpc_id
+  target_type          = "ip"
+  deregistration_delay = var.deregistration_delay
 
   health_check {
     port                = var.http_probe_port
@@ -15,6 +16,18 @@ resource "aws_lb_target_group" "gwlb_target_group" {
     interval            = var.health_check_interval
     healthy_threshold   = var.healthy_threshold
     unhealthy_threshold = var.unhealthy_threshold
+  }
+
+  target_failover {
+    on_deregistration = var.rebalance_enabled == true ? "rebalance" : "no_rebalance"
+    on_unhealthy      = var.rebalance_enabled == true ? "rebalance" : "no_rebalance"
+  }
+
+  # type attribute only applies if enabled = true and only options are "source_ip_dest_ip" (2-tuple) or "source_ip_dest_ip_proto" (3-tuple).
+  # enabled = false implies 5-tuple. AWS gives type a default value of "source_ip_dest_ip_proto" even if enabled is set to false
+  stickiness {
+    enabled = var.flow_stickiness == "5-tuple" ? false : true
+    type    = var.flow_stickiness == "2-tuple" ? "source_ip_dest_ip" : "source_ip_dest_ip_proto"
   }
 }
 
@@ -76,13 +89,13 @@ resource "aws_lb_target_group_attachment" "gwlb_target_group_attachment_lrg_3" {
 ################################################################################
 resource "aws_lb" "gwlb" {
   load_balancer_type               = "gateway"
-  name                             = "${var.name_prefix}-cc-gwlb-${var.resource_tag}"
+  name                             = var.gwlb_name
   enable_cross_zone_load_balancing = var.cross_zone_lb_enabled
 
   subnets = var.cc_subnet_ids
 
   tags = merge(var.global_tags,
-    { Name = "${var.name_prefix}-gwlb-${var.resource_tag}" }
+    { Name = var.gwlb_name }
   )
 }