+++ modules/terraform-zscc-iam-aws/versions.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 4.7.0"
+      version = "~> 4.59.0"
     }
   }
   required_version = ">= 0.13.7, < 2.0.0"